Khi add product vào cart, check số lượng (> 0) cả FE lẫn BE Sửa lại luồng đặt hàng  đã giao thì phải chuyển sang đã thanh toán Sau khi add-to-cart vẫn ở trang hiện tại

+++ Do an TMDT/ViewModels/itemcart.cs
@@ -11,6 +11,6 @@ namespace Do_an_CCNPMM.ViewModels
         public ChiTietGioHang CT_GH { get; set; }
         public MatHang SanPham { get; set; }
         public List<MatHangAnh> MatHangAnhs { get; set; }
-        public int tong { get; set; }
+        public long tong { get; set; }
     }
 }
